requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
python-dotenv==1.0.1
typing_extensions==4.12.1
yarl==1.9.4
